Added a helper file for CarList.tsx
